feat: refactor aggregator with pluggable strategies
