+++ flsim0/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["config","storage","models","partition","node","aggregator","eval","attacks","runner"]
+__all__ = ["config","storage","models","partition","node","aggregator","aggregation","eval","attacks","runner"]
