+++ flsim0/config.py
@@ -59,6 +59,10 @@ class PathsCfg:
     log_csv: Path = Path("./runs/tmp/fl_log.csv")
 
 @dataclass
+class AggregationCfg:
+    strategy: str = "fedavg"
+
+@dataclass
 class Cfg:
     data: DataCfg
     model: ModelCfg            # NEW
@@ -66,6 +70,7 @@ class Cfg:
     reward: RewardCfg
     attack: AttackCfg
     paths: PathsCfg
+    aggregation: AggregationCfg
 
 def build_config(args, run_dir: Path):
     paths = PathsCfg(
@@ -81,6 +86,7 @@ def build_config(args, run_dir: Path):
     train = TrainCfg()
     reward = RewardCfg()
     attack = AttackCfg()
+    aggregation = AggregationCfg()
 
     # ---------- overrides: data ----------
     if getattr(args, "dataset", None) is not None:    # NEW: --dataset
@@ -130,4 +136,6 @@ def build_config(args, run_dir: Path):
     if getattr(args, "spoof_mode", None) is not None:          attack.spoof_mode = args.spoof_mode
     if getattr(args, "malicious_seed", None) is not None:      attack.seed = args.malicious_seed
 
-    return Cfg(data=data, model=model, train=train, reward=reward, attack=attack, paths=paths)
+    if getattr(args, "agg_strategy", None) is not None:        aggregation.strategy = args.agg_strategy
+
+    return Cfg(data=data, model=model, train=train, reward=reward, attack=attack, paths=paths, aggregation=aggregation)
